Reorganized Function & FunctionBuilder
